User_login: Refactoring
